--- src/main/java/com/example/demo/domain/board/service/component/command/ViewIncreaseCommand.java
@@ -1,9 +1,0 @@
-package com.example.demo.domain.board.service.component.command;
-
-import com.example.demo.domain.board.Repository.BoardRepository;
-import com.example.demo.domain.board.Repository.ViewRepository;
-
-
-public interface ViewIncreaseCommand {
-    void increaseView(Long boardId);
-}
