+++ src/main/java/com/example/demo/domain/board/service/service/ViewIncreaseService.java
@@ -1,0 +1,5 @@
+package com.example.demo.domain.board.service.service;
+
+public interface ViewIncreaseService {
+    void increaseView(Long boardId);
+}
